Nth commit

+++ t/90-Analysis-Plot.t
@@ -4,11 +4,11 @@ use 5.006;
 use strict;
 use warnings;
 
-our $VERSION = '0.21';
+our $VERSION = '0.25';
 
 use lib 'blib/lib';
 
-use Statistics::Covid::Datum::IO;
+use Statistics::Covid;
 use Statistics::Covid::Analysis::Plot::Simple;
 use Test::More;
 use File::Basename;
@@ -18,6 +18,10 @@ use File::Path;
 
 use Data::Dump qw/pp/;
 
+my $delete_out_files = 0;
+my $DEBUG = 0;
+
+### nothing to change below
 my $dirname = dirname(__FILE__);
 
 my $num_tests = 0;
@@ -37,23 +41,34 @@ my $dbfullpath = File::Spec->catfile($confighash->{'dbparams'}->{'dbdir'}, $conf
 
 ok(-f $dbfullpath, "found test db"); $num_tests++;
 
-my $io = Statistics::Covid::Datum::IO->new({
+my $covid = Statistics::Covid->new({
 	'config-hash' => $confighash,
-	'debug' => 0,
+	'debug' => $DEBUG,
 });
+ok(defined($covid), "Statistics::Covid::Datum::IO->new() called"); $num_tests++;
+ok(defined($covid->datum_io()), "connect to db: '$dbfullpath'."); $num_tests++;
 
-ok(defined($io), "Statistics::Covid::Datum::IO->new() called"); $num_tests++;
-ok($io->db_connect(), "connect to db: '$dbfullpath'."); $num_tests++;
+ok(exists($confighash->{'analysis'}), "analysis sub-key in config hash exists") or BAIL_OUT; $num_tests++;
+ok(exists($confighash->{'analysis'}->{'plot'}), "analysis/plot sub-key in config hash exists") or BAIL_OUT; $num_tests++;
+ok(defined($confighash->{'analysis'}->{'plot'}), "analysis/plot sub-key in confighash defined") or BAIL_OUT; $num_tests++;
 
-my $objs = $io->db_select({
-	conditions => {belongsto=>'UK', name=>{'like' => 'Ha%'}}
+my $plotparams = $confighash->{'analysis'}->{'plot'};
+$plotparams->{'debug'} = $DEBUG;
+
+# get datum objs for specific location sorted wrt time
+my $objs = $covid->select_datums_from_db_time_ascending({
+	'conditions' => {
+		'admin3' => {'like' => 'Hack%'},
+		#or {'like' => 'Haver%'},
+		'admin0' => 'United Kingdom of Great Britain and Northern Ireland'
+	}
 });
-ok(defined($objs), "db_select() called.") || BAIL_OUT("can not continue, something wrong with the test db which should have been present in t dir"); $num_tests++;
-ok(scalar(@$objs)>0, "db_select() returned objects.") || BAIL_OUT("can not continue, something wrong with the test db which should have been present in t dir"); $num_tests++;
+ok(defined($objs), "select_datums_from_db_time_ascending() called.") || BAIL_OUT("can not continue, something wrong with the test db which should have been present in t dir"); $num_tests++;
+ok(scalar(@$objs)>0, "select_datums_from_db_time_ascending() returned objects.") || BAIL_OUT("can not continue, something wrong with the test db which should have been present in t dir"); $num_tests++;
 
 my $df = Statistics::Covid::Utils::datums2dataframe({
 	'datum-objs' => $objs,
-	'groupby' => ['name','belongsto'],
+	'groupby' => ['admin0', 'admin3'],
 	'content' => ['confirmed','unconfirmed','datetimeUnixEpoch'],
 });
 ok(defined($df), "Statistics::Covid::Utils::datums2dataframe() called."); $num_tests++;
@@ -68,9 +83,10 @@ $ret = Statistics::Covid::Analysis::Plot::Simple::plot({
 	'outfile' => $outfile,
 	'Y' => 'confirmed',
 	'date-format-x' => 123,
-	'GroupBy' => ['name']
+	'GroupBy' => ['admin1'],
+	%$plotparams
 });
-ok(!defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called"); $num_tests++;
+ok(!defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called to fail."); $num_tests++;
 # fail because no dataframe or datum-objs
 $ret = Statistics::Covid::Analysis::Plot::Simple::plot({
 	'outfile' => $outfile,
@@ -80,26 +96,12 @@ $ret = Statistics::Covid::Analysis::Plot::Simple::plot({
 		position => 'bottom',
 		orientation => 'horizontal'
 	},
-	'GroupBy' => ['name']
+	'GroupBy' => ['admin1'],
+	%$plotparams
 });
-ok(!defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called"); $num_tests++;
+ok(!defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called to fail."); $num_tests++;
 
-# call with dataframe instead
-$ret = Statistics::Covid::Analysis::Plot::Simple::plot({
-	'dataframe' => $df,
-	'outfile' => $outfile,
-	'Y' => 'confirmed',
-	'date-format-x' => {
-		format => '%d/%m',
-		position => 'bottom',
-		orientation => 'horizontal'
-	},
-});
-ok(defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called with dataframe"); $num_tests++;
-ok((-f $outfile)&&(-s $outfile), "output image '$outfile'."); $num_tests++;
-unlink $outfile;
-
-# success, call with datum-objs
+# fail, call with datum-objs which is no longer supported
 $ret = Statistics::Covid::Analysis::Plot::Simple::plot({
 	'datum-objs' => $objs,
 	'outfile' => $outfile,
@@ -110,21 +112,39 @@ $ret = Statistics::Covid::Analysis::Plot::Simple::plot({
 		position => 'bottom',
 		orientation => 'horizontal'
 	},
-	'GroupBy' => ['name']
+	'GroupBy' => ['admin1'],
+	%$plotparams
 });
-ok(defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called with datum-objs"); $num_tests++;
-ok((-f $outfile)&&(-s $outfile), "output image '$outfile'."); $num_tests++;
-unlink $outfile;
+ok(!defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called with datum-objs"); $num_tests++;
 
-# call with dataframe, X is not time
+# succeed, call with dataframe, X is not time
+$ret = Statistics::Covid::Analysis::Plot::Simple::plot({
+	'dataframe' => $df,
+	'outfile' => $outfile,
+	'title' => 'test-title',
+	'with-legend' => 0,
+	'Y' => 'confirmed',
+	'X' => 'confirmed', # because of failures in case of no variation
+	%$plotparams
+});
+ok(defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called with dataframe"); $num_tests++;
+ok((-f $outfile)&&(-s $outfile), "output image '$outfile'."); $num_tests++;
+unlink $outfile if $delete_out_files; 
+
+# succeed, call with dataframe instead
 $ret = Statistics::Covid::Analysis::Plot::Simple::plot({
 	'dataframe' => $df,
 	'outfile' => $outfile,
 	'Y' => 'confirmed',
-	'X' => 'unconfirmed',
+	'date-format-x' => {
+		format => '%d/%m',
+		position => 'bottom',
+		orientation => 'horizontal'
+	},
+	%$plotparams
 });
 ok(defined($ret), "Statistics::Covid::Analysis::Plot::Simple::plot() called with dataframe"); $num_tests++;
 ok((-f $outfile)&&(-s $outfile), "output image '$outfile'."); $num_tests++;
-unlink $outfile;
+unlink $outfile if $delete_out_files;
 
 done_testing($num_tests);
